requirements.txt
flask==3.0.3
flask-cors==4.0.1
web3==6.20.0
requests==2.32.3
python-dotenv==1.0.1
gunicorn==22.0.0
py-solc-x==2.1.1
